+++ staticsite/site.py
@@ -215,24 +215,6 @@ It defaults to true at least for [Markdown](markdown.md),
         from .archetypes import Archetypes
         return Archetypes(self, os.path.join(self.settings.PROJECT_ROOT, "archetypes"))
 
-    def find_theme_root(self) -> str:
-        """
-        Choose a theme root from the ones listed in the configuration
-        """
-        # Pick the first valid theme directory
-        candidate_themes = self.settings.THEME
-        if isinstance(candidate_themes, str):
-            candidate_themes = (candidate_themes,)
-
-        for theme_root in candidate_themes:
-            theme_root = os.path.join(self.settings.PROJECT_ROOT, theme_root)
-            if os.path.isdir(theme_root):
-                return theme_root
-
-        raise RuntimeError(
-                "None of the configured theme directories ({}) seem to exist".format(
-                    ", ".join(self.settings.THEME)))
-
     def load_theme(self):
         """
         Load a theme from the given directory.
@@ -240,14 +222,22 @@ It defaults to true at least for [Markdown](markdown.md),
         This needs to be called once (and only once) before analyze() is
         called.
         """
+        from .theme import Theme
+
         if self.theme is not None:
             raise RuntimeError(
                     F"load_theme called while a theme was already loaded from {self.theme.root}")
 
-        theme_root = self.find_theme_root()
+        if isinstance(self.settings.THEME, str):
+            self.theme = Theme.create(self, self.settings.THEME)
+        else:
+            # Pick the first valid theme directory
+            candidate_themes = self.settings.THEME
+            if isinstance(candidate_themes, str):
+                candidate_themes = (candidate_themes,)
+            self.theme = Theme.create_legacy(self, candidate_themes)
 
-        from .theme import Theme
-        self.theme = Theme(self, theme_root)
+        self.theme.load()
 
     def _settings_to_meta(self) -> Meta:
         """