+++ tests/test_templates.py
@@ -17,7 +17,8 @@ class MockPage:
 class TestTemplates(TestCase):
     def test_arrange(self):
         site = test_utils.Site()
-        theme = Theme(site, ".")
+        theme = Theme(site, "test", [{"root": ".", "name": "test"}])
+        theme.load()
 
         self.maxDiff = None